+++ src/Kit/Navigation/Header/Header.tsx
@@ -5,7 +5,7 @@ import cn from 'classnames';
 import styles from './styles.module.scss';
 import {KitDropDown} from '../DropDown';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {faBars} from '@fortawesome/free-solid-svg-icons';
+import {faGrip} from '@fortawesome/free-solid-svg-icons';
 import {getContrastColor} from '@utils/functions';
 import {KitColorProp} from '@utils/functions/types';
 
@@ -36,7 +36,7 @@ export const Header: FunctionComponent<IKitHeader> = ({
         className
     );
 
-    const _renderDropdown = () => menu;
+    const _renderDropdown = () => <div className="kit-header-menu-wrapper">{menu}</div>;
 
     const customColors = _getCustomColors(theme.colors.primary['400']);
 
@@ -44,8 +44,8 @@ export const Header: FunctionComponent<IKitHeader> = ({
         <div style={customColors} {...props} className={clx}>
             {menu && (
                 <KitDropDown dropdownRender={_renderDropdown}>
-                    <div className="kit-header-menu">
-                        <FontAwesomeIcon icon={faBars} />
+                    <div className="kit-header-menu" tabIndex={0}>
+                        <FontAwesomeIcon icon={faGrip} />
                     </div>
                 </KitDropDown>
             )}