+++ stories/02-Design System/03-Navigation/Header/commons.tsx
@@ -1,11 +1,9 @@
-import {KitIdCard} from '@kit/DataDisplay';
 import {CSSProperties} from 'react';
-import {KitSelect, KitSpace, KitTypography} from '@kit/index';
+import {KitButton, KitSelect, KitSpace, KitTypography} from '@kit/index';
 import {FlagIcon} from 'react-flag-kit';
 import {KitHeader} from '@kit/Navigation';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {faUser} from '@fortawesome/free-regular-svg-icons';
-import {faRightFromBracket} from '@fortawesome/free-solid-svg-icons';
+import {faMasksTheater, faRightFromBracket} from '@fortawesome/free-solid-svg-icons';
 
 const flagWidth = 20;
 
@@ -48,15 +46,6 @@ export const profileMenuContentExemple = {
             key: '1',
             label: (
                 <KitSpace size={16}>
-                    <FontAwesomeIcon icon={faUser} /> Mon compte
-                    <KitTypography.Text></KitTypography.Text>
-                </KitSpace>
-            )
-        },
-        {
-            key: '2',
-            label: (
-                <KitSpace size={16}>
                     <FontAwesomeIcon icon={faRightFromBracket} />
                     <KitTypography.Text>Deconnexion</KitTypography.Text>
                 </KitSpace>
@@ -69,11 +58,47 @@ export const logoExample = <img src="public/images/omnipublish-logo.svg" alt="" 
 
 export const userProfileExample = (
     <KitHeader.Profile
+        data-testid="profile"
         logo="/public/images/logo.png"
         alt="Logo ARiSTiD"
         tagProps={{type: 'primary', idCardProps: {description: 'Grenoble Centre'}}}
         menu={profileMenuContentExemple}
-        userCard={<KitIdCard avatarProps={{src: 'public/images/portrait.png'}} />}
+        profileMenuCardProps={{
+            avatarProps: {
+                src: 'public/images/portrait.png'
+            },
+            title: 'Nina',
+            description: 'Responsable Marketing',
+            onClick: () => console.log('profile menu card clicked')
+        }}
+        profileCardProps={{
+            avatarProps: {
+                src: 'public/images/portrait.png'
+            }
+        }}
+    />
+);
+
+export const userProfileComplementExample = (
+    <KitHeader.Profile
+        logo="/public/images/logo.png"
+        alt="Logo ARiSTiD"
+        tagProps={{type: 'primary', idCardProps: {description: 'Grenoble Centre'}}}
+        menu={profileMenuContentExemple}
+        profileMenuCardProps={{
+            avatarProps: {
+                src: 'public/images/portrait.png'
+            },
+            title: 'Nina',
+            description: 'Responsable Marketing',
+            onClick: () => console.log('profile menu card clicked')
+        }}
+        profileCardProps={{
+            avatarProps: {
+                src: 'public/images/portrait.png'
+            }
+        }}
+        complement={<KitButton icon={<FontAwesomeIcon icon={faMasksTheater} />} />}
     />
 );