+++ stories/02-Design System/03-Navigation/Header/examples/index.tsx
@@ -3,6 +3,7 @@ import Basic from './basic';
 import Lang from './lang';
 import Menu from './menu';
 import Profile from './profile';
+import ProfileComplement from './profileComplement';
 import Logo from './logo';
 
 import AllSource from './all?raw';
@@ -10,6 +11,7 @@ import BasicSource from './basic?raw';
 import LangSource from './lang?raw';
 import MenuSource from './menu?raw';
 import ProfileSource from './profile?raw';
+import ProfileComplementSource from './profileComplement?raw';
 import LogoSource from './logo?raw';
 
 export const Sources = {
@@ -18,6 +20,7 @@ export const Sources = {
     Lang: LangSource,
     Menu: MenuSource,
     Profile: ProfileSource,
+    ProfileComplement: ProfileComplementSource,
     Logo: LogoSource
 };
 
@@ -27,5 +30,6 @@ export default {
     Lang,
     Menu,
     Profile,
+    ProfileComplement,
     Logo
 };